Use fsharpc for a proper compile
Before, we were just inserting a call to
"Environment.Exit(0)" at the beginning of
the script, and calling it with fsharpi,
which would catch only syntax errors.

But this commit now introduces a real
fsharpc compilation which will catch
semantic errors as well.

The testcase to determine that this change
indeed works, has been trying to compile
fsx itself.