Honour flags properly
Before this commit, fsx was just compiling the file and nothing else.
Now fsx, by default, compiles the file and runs it (calling mono).

If the -c flag is specified, it only compiles the script and places
in the bin/ folder.

If the -k flag is specified, it only chec*k*s if the file builds,
which means that it compiles it but removes the binary.

The changes to Infra.fs in this commit are because fsx needs to
show the output of the compilation, or not, depending on the final
exitCode, so then it's better to store all output in the same order
it arrives (via the events), so that we can print it later in the
same order.