InfraLib/MiscTools: fix FsxArguments detection on Windows
The .LocalPath property of Uri class was truncating the path
when it found a # character, and given that Fsi.exe lives in
some Windows folder that has 'F#' in its path, it was getting
an incomplete path, and not being able to recognize Fsi.exe
as not an argument to the script (also due to comparing in
a case sensitive way, which is also fixed in this commit).

The solution is rather ugly, but it's the only one that works
that I could find. See: https://stackoverflow.com/a/41203269