CI: Snap build on Ubuntu 20.04
Snap/Snapcraft does not run properly on the default Ubuntu 20.04 image
used by Gitlab. This commit implements a docker-in-docker (DinD) container
running a specialized configuration for snap and snapcraft which is executed
inside the default Gitlab Ubuntu 20.04 container.
The 18.04 snap build lane is removed.

Blog post for more information on the snap/snapcraft container:
https://ograblog.wordpress.com/2017/06/02/dock-a-snap/