InfaLib/Process.fs: don't cross the streams, Venkman!
The way the lock() statement was placed allowed for characters from
StdOut and StdErr to be constantly intermixed by the process owner;
but if we wait until the `\n` arrives and exit the lock only then,
it lets the streams to not be interleaved with each other.

This is similar to this previous attempt [1] (reverted in [2]),
however it seems that the QueuedLock I borrowed [3] was implemented
incorrectly (this time I followed this advice to use volatile
in F#: [4]), and it was a convoluted anyway, I like this simpler
approach more but I essentially implemented the same solution in a
different way (it's still not perfect because lines still
get interleaved, but at least not single characters).

(Concurrency tests still pass.)

[1] 0c691df6a76c6e31d0b5fc15216c21075d1c35aa
[2] c47382f0af602081c99e0c7548ab2d32f1755ca8
[3] https://stackoverflow.com/a/961904/544947
[4] https://stackoverflow.com/a/38590178/544947