Fix fsxc build when using new mono with old F#
In this case we suffer a similar problem as the one
described in [1]. However, the fix here doesn't seem
to need the use of xbuild, but copying the same
Microsoft.FSharp.Targets finding strategy that we
already have in InfraLib.fsproj. This may have some
disadvantages (such as failing to embed files or
bad stacktraces that don't include file names even
in debug mode), but we're not hit by those yet, and
this fix is good enough for now.

[1] https://github.com/nblockchain/geewallet/commit/9d76c300687450c4d737338930f7b2559a80eaa9