Windows: prevent crash due to not finding FSharp.Core.dll
fsxc.exe was being copied to %ProgramFiles% properly thanks to
the previous commit, but it could not run because it depended
on FSharp.Core.dll and so was crashing.

Unlike with Unix systems, we copy the DLL from nuget in Windows,
to not to make fsx rely even more than it does on .NET-legacy
infrastructure. When we move to .NET6 soon we can even remove
the download of nuget.exe (use 'dotnet restore' instead) and the
dependency on xbuild/MSBuild (and use 'dotnet build' instead).