InfraLib/Process: nicer API than Exec vs SafeExec
It was always better to use SafeExec API (because if exitCode
was diff than zero it would raise an exception), however: given
that it had a different overload, it would be easy to miss.
This should have been the default API, and have the non-Safe one
be called UnsafeExec; but just overhauling the result type of
this is a better approach.