Windows: fix fsx.bat to propagate errorlevel
The previous commit[1] failed CI but only thanks to our recently
introduced Warning detection (included in our recent Process API
change[2]). This way we're certain the errorlevel was not properly
propagated in Windows because it should have been different than 0
while only a warning was detected (exitCode=0 but StdErr.Length>0).

[1] 1dc56edc9f08628adb4d5da4c976efeb1d040254
[2] f47331f58709f0c0953ced2d6538cb996f8df49f