Disable GitLabCI
GitLab.com is a joke.  First they promised free CI for everyone.
Then they cut it down to only public (opensource) projects.
Then they reduced it to 2,000 minutes per month.  Then they
reduced it even more and on the 4th day of the month we have
already ran out of minutes.

Well, we then now move to GitHubCI that has unlimited minutes
for opensource repos, plus it's cross-platform (not like
GitLab.com that only has Linux runners) and also has a better
step system that doesn't need to pass artifacts around the
pipelines all the time.

For a bit of background on how we actually liked and supported
GitLab in the early days, read this other commit message[1] from
a sibling repository.

The ice on the cake for this commit is the addition of a
GitHubActions YML file sample to the ReadMe.md file, alongside
the GitLabCI one.

[1] https://github.com/nblockchain/geewallet/commit/0a3b3448f439117e41ae527bbbf5ac54b4199907