Greatly simplify source autogeneration
Before this commit, we used to have this process:
1. fsxc instructed to compile foo/bar.fsx.
2. foo/bin/ dir created.
3. foo/bin/bar.fsx created, with necessary modifications to be
compiled.
4. foo/bar.fsx.bak created with original contents of
foo/bar.fsx.
5. foo/bar.fsx overwritten with content from foo/bin/bar.fsx.
6. Invoke F# compiler.
7. Restore foo/bar.fsx from backup foo/bar.fsx.bak.

This above is such a stupid and complicated process, when we can
just create foo/bin/bar.fsx.fs in step 3, and feed that to the
compiler.  This way we don't need to keep backups and restore
from backup later.

And now, if a user hits CTRL+C when in the middle of an fsxc
compilation, she will not find their .fsx script modified.