InfraLib/Unix.fs: implement SudoIfNeeded() & prefer it
Sudo() call fails if user is already root, so this version
instead just runs the command when this happens.  This is better
for scenarios in which we may want some code to be the same
across vanilla&normal VMs (so both CIs run properly even if one
of them is already running with privileges).  For this reason,
we replace all usages of Sudo() in other functions with this new
one.