fsxc: de-prioritize netstandard2.0
The recently implemented `#r "NugetPkg"` feature doesn't really
work for net4.x legacy when trying to consume netstandard2.0
binaries because it generates this error:

```
error FS0078: Unable to find the file 'netstandard.dll' in any of...
```

The above is demonstrated by the previous commit[1]'s CI job.

This commit doesn't fix the problem completely but mitigates it.
At least it will make it work for nuget packages that also
distribute net4.x binaries (and while we're at it, we add net452
which the TickSpec (the lib used in our nugetRef tests now)
uses.

[1] 0c2ecbd2d1f420aa924fba31d953b3e4e80d7f6c