launcher.sh,fsx: stop crash when --help is used
If someone used --help on fsxc, it worked fine, but not
if someone used it on fsx: on Windows it would crash saying
fsx.exe/fsx.dll was not found, and on Unix it would exit with
a message saying "please report this bug" and exitCode<>0.

This "please report this bug" msg was missing from fsx's
Program.fs, so we added it there & fixed the crash on Windows.
And interestingly enough, to fix the problem in Unix, we had
to remove the message from there and exit 0 directly (in the
Windows case, as it's using fsx's Program.fs which calls
fsxc's Program.fs under the hood, then it reaches an
Environment.Exit invocation there so that case doesn't happen
and the message "please report this bug" can stay).

Side-note, I used this abbreviation below to shorten a comment:
https://github.com/nblockchain/conventions/commit/ee7aa57de76e7721f2af69cbb637b09b6fb25d08