Fsdk/Misc: fix NRE if legacy
When running TSV unit tests in legacy framework (e.g. Mono),
a NRE was being thrown; I guess because in that case the
function GetEntryAssembly() was returning null.

This fix is kinda ugly but is anyway affecting a part of the
code that I plan to retire soon (cause it's API wrapped under
LEGACY_FRAMEWORK ifdef).