Fsdk(,.Tests): optional prog pred (failing test)
When using this in a program (as opposed to a .fsx script),
there's no need to use fsi.CommandLineArgs or
Environment.GetCommandLineArgs(), because we just can just use
"argv" (an array<string>) from the Main method, so in that case
we don't need the _prog_ram name _pred_icate, as it's already
stripped.

So let's morph all tests to also consider this way, which will
make them all fail, but we try to fix everything on the next
commit after this.