fsx,fsxc: deprecate --check/-k flag
In the cases where launcher.sh is used (instead of fsx.fsproj),
the -k flag was being passed to fsxc successfully, but then the
script would be run. As fixing this would be a bit messy (would
need lots of bash hacks in the launcher), let's rather
deprecate it; and if someone needs to only compile the script,
they can call fsxc directly (which is what we advice in the
deprecation error message).

Fixes https://github.com/nblockchain/fsx/issues/37