scripts/publish: prevent uploading old versions
After several 0.6.0--date... versions were published,
I pushed the 0.6.0 tag and that was it. Then after
that I changed the version.config to make BaseVersion
be 0.6.1 and published a 0.6.1--date... because of this.
However, I had to revert the commit and I forgot to not
include the BaseVersion change in the revert, so it
meant that a new 0.6.0--date... prerelease was published
even after v0.6.0 had already been published!

With this check in the publish script we will prevent this
from happening again.